+++ PriceAggregatorAPI/Models/DTOs/CategoryDTO.cs
@@ -2,7 +2,7 @@ namespace PriceAggregatorAPI.Models.DTOs;
 
 public class CategoryDto
 {
-    private int Id { get; set; }
+    private long Id { get; set; }
     public string Name { get; set; }
     public int TreeLevel { get; set; }
     public int ParentId { get; set; }
